requirements.txt
flask
python-telegram-bot==20.7
cloudscraper
beautifulsoup4